Start refactoring the code and unit testing it.
Code now is divided in:

- gauss.c -> The main function; is supposed to give the
functionality that the program wants.
- gauss_func.c -> Functions that are essential for the program
execution.
- test.c -> Unit tests of functions contained in "gauss_func.c"
powered by the microframework "minunit.h".
